moving viewmodelbase to usercontrol (will eventually be removed)
